+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.1.2 under development
 
-- no changes in this release.
+- Chg #44: Bump minimum PHP version to 8.1 (@vjik) 
+- Chg #44: Change PHP constraint in composer.json to `8.1 - 8.4` (@vjik)
+- Enh #44: Add psalm type `int<1, 2147483647>` to `depth` parameter in `JsonParser` constructor (@vjik)
+- Bug #44: Explicitly mark nullable parameters (@vjik)
 
 ## 1.1.1 June 03, 2024